chore(firestore): prep for firestore pipelines GA (#16549)
Misc changes in prep for GA release of pipelines:

- re-enable releases for firestore
- remove notes about pipelines being in preview
    - added individual beta labels for search/dml classes and methods
- expose options in raw_stage
- renamed Type to PipelineDataType
- moved helper classes (options, enums) into pipeline_types
- expose pipeline classes in top-level \_\_init__.py
