+++ packages/google-cloud-firestore/google/cloud/firestore/__init__.py
@@ -30,6 +30,8 @@ from google.cloud.firestore_v1 import (
     AsyncClient,
     AsyncCollectionReference,
     AsyncDocumentReference,
+    AsyncPipeline,
+    AsyncPipelineStream,
     AsyncQuery,
     AsyncTransaction,
     AsyncWriteBatch,
@@ -43,15 +45,31 @@ from google.cloud.firestore_v1 import (
     ExistsOption,
     ExplainOptions,
     FieldFilter,
+    FindNearestOptions,
     GeoPoint,
     Increment,
     LastUpdateOption,
     Maximum,
     Minimum,
     Or,
+    Ordering,
+    Pipeline,
+    PipelineDataType,
+    PipelineExplainOptions,
+    PipelineResult,
+    PipelineSnapshot,
+    PipelineSource,
+    PipelineStream,
     Query,
     ReadAfterWriteError,
+    SampleOptions,
+    SearchOptions,
+    SubPipeline,
+    TimeGranularity,
+    TimePart,
+    TimeUnit,
     Transaction,
+    UnnestOptions,
     Watch,
     WriteBatch,
     WriteOption,
@@ -68,6 +86,8 @@ __all__: List[str] = [
     "AsyncClient",
     "AsyncCollectionReference",
     "AsyncDocumentReference",
+    "AsyncPipeline",
+    "AsyncPipelineStream",
     "AsyncQuery",
     "async_transactional",
     "AsyncTransaction",
@@ -83,18 +103,34 @@ __all__: List[str] = [
     "ExistsOption",
     "ExplainOptions",
     "FieldFilter",
+    "FindNearestOptions",
     "GeoPoint",
     "Increment",
     "LastUpdateOption",
     "Maximum",
     "Minimum",
     "Or",
+    "Ordering",
+    "Pipeline",
+    "PipelineDataType",
+    "PipelineExplainOptions",
+    "PipelineResult",
+    "PipelineSnapshot",
+    "PipelineSource",
+    "PipelineStream",
     "Query",
     "ReadAfterWriteError",
     "SERVER_TIMESTAMP",
+    "SampleOptions",
+    "SearchOptions",
+    "SubPipeline",
+    "TimeGranularity",
+    "TimePart",
+    "TimeUnit",
     "Transaction",
     "transactional",
     "types",
+    "UnnestOptions",
     "Watch",
     "WriteBatch",
     "WriteOption",
