+++ packages/google-cloud-firestore/google/cloud/firestore_v1/async_pipeline.py
@@ -11,11 +11,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""
-.. warning::
-    **Preview API**: Firestore Pipelines is currently in preview and is
-    subject to potential breaking changes in future releases
-"""
 
 from __future__ import annotations
 
@@ -61,9 +56,7 @@ class AsyncPipeline(_BasePipeline):
 
     Use `client.pipeline()` to create instances of this class.
 
-    .. warning::
-        **Preview API**: Firestore Pipelines is currently in preview and is
-        subject to potential breaking changes in future releases
+
     """
 
     _client: AsyncClient
