+++ packages/google-cloud-firestore/google/cloud/firestore_v1/pipeline_expressions.py
@@ -11,11 +11,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""
-.. warning::
-    **Preview API**: Firestore Pipelines is currently in preview and is
-    subject to potential breaking changes in future releases.
-"""
 
 from __future__ import annotations
 
@@ -35,8 +30,15 @@ if TYPE_CHECKING:
     from google.cloud.firestore_v1.base_pipeline import _BasePipeline
 
 from google.cloud.firestore_v1._helpers import GeoPoint, decode_value, encode_value
-from google.cloud.firestore_v1.types.document import Value
+from google.cloud.firestore_v1.pipeline_types import (
+    Ordering,
+    PipelineDataType,
+    TimeGranularity,
+    TimePart,
+    TimeUnit,
+)
 from google.cloud.firestore_v1.types.document import Pipeline as Pipeline_pb
+from google.cloud.firestore_v1.types.document import Value
 from google.cloud.firestore_v1.types.query import StructuredQuery as Query_pb
 from google.cloud.firestore_v1.vector import Vector
 
@@ -52,142 +54,6 @@ CONSTANT_TYPE = TypeVar(
     Vector,
     None,
 )
-
-
-class TimeUnit(str, Enum):
-    """Enumeration of the different time units supported by the Firestore backend."""
-
-    MICROSECOND = "microsecond"
-    MILLISECOND = "millisecond"
-    SECOND = "second"
-    MINUTE = "minute"
-    HOUR = "hour"
-    DAY = "day"
-
-
-class TimeGranularity(str, Enum):
-    """Enumeration of the different time granularities supported by the Firestore backend."""
-
-    # Inherit from TimeUnit
-    MICROSECOND = TimeUnit.MICROSECOND.value
-    MILLISECOND = TimeUnit.MILLISECOND.value
-    SECOND = TimeUnit.SECOND.value
-    MINUTE = TimeUnit.MINUTE.value
-    HOUR = TimeUnit.HOUR.value
-    DAY = TimeUnit.DAY.value
-
-    # Additional granularities
-    WEEK = "week"
-    WEEK_MONDAY = "week(monday)"
-    WEEK_TUESDAY = "week(tuesday)"
-    WEEK_WEDNESDAY = "week(wednesday)"
-    WEEK_THURSDAY = "week(thursday)"
-    WEEK_FRIDAY = "week(friday)"
-    WEEK_SATURDAY = "week(saturday)"
-    WEEK_SUNDAY = "week(sunday)"
-    ISOWEEK = "isoweek"
-    MONTH = "month"
-    QUARTER = "quarter"
-    YEAR = "year"
-    ISOYEAR = "isoyear"
-
-
-class TimePart(str, Enum):
-    """Enumeration of the different time parts supported by the Firestore backend."""
-
-    # Inherit from TimeUnit
-    MICROSECOND = TimeUnit.MICROSECOND.value
-    MILLISECOND = TimeUnit.MILLISECOND.value
-    SECOND = TimeUnit.SECOND.value
-    MINUTE = TimeUnit.MINUTE.value
-    HOUR = TimeUnit.HOUR.value
-    DAY = TimeUnit.DAY.value
-
-    # Inherit from TimeGranularity
-    WEEK = TimeGranularity.WEEK.value
-    WEEK_MONDAY = TimeGranularity.WEEK_MONDAY.value
-    WEEK_TUESDAY = TimeGranularity.WEEK_TUESDAY.value
-    WEEK_WEDNESDAY = TimeGranularity.WEEK_WEDNESDAY.value
-    WEEK_THURSDAY = TimeGranularity.WEEK_THURSDAY.value
-    WEEK_FRIDAY = TimeGranularity.WEEK_FRIDAY.value
-    WEEK_SATURDAY = TimeGranularity.WEEK_SATURDAY.value
-    WEEK_SUNDAY = TimeGranularity.WEEK_SUNDAY.value
-    ISOWEEK = TimeGranularity.ISOWEEK.value
-    MONTH = TimeGranularity.MONTH.value
-    QUARTER = TimeGranularity.QUARTER.value
-    YEAR = TimeGranularity.YEAR.value
-    ISOYEAR = TimeGranularity.ISOYEAR.value
-
-    # Additional parts
-    DAY_OF_WEEK = "dayofweek"
-    DAY_OF_YEAR = "dayofyear"
-
-
-class Ordering:
-    """Represents the direction for sorting results in a pipeline."""
-
-    class Direction(Enum):
-        ASCENDING = "ascending"
-        DESCENDING = "descending"
-
-    def __init__(self, expr, order_dir: Direction | str = Direction.ASCENDING):
-        """
-        Initializes an Ordering instance
-
-        Args:
-            expr (Expression | str): The expression or field path string to sort by.
-                If a string is provided, it's treated as a field path.
-            order_dir (Direction | str): The direction to sort in.
-                Defaults to ascending
-        """
-        self.expr = expr if isinstance(expr, Expression) else Field.of(expr)
-        self.order_dir = (
-            Ordering.Direction[order_dir.upper()]
-            if isinstance(order_dir, str)
-            else order_dir
-        )
-
-    def __repr__(self):
-        if self.order_dir is Ordering.Direction.ASCENDING:
-            order_str = ".ascending()"
-        else:
-            order_str = ".descending()"
-        return f"{self.expr!r}{order_str}"
-
-    def _to_pb(self) -> Value:
-        return Value(
-            map_value={
-                "fields": {
-                    "direction": Value(string_value=self.order_dir.value),
-                    "expression": self.expr._to_pb(),
-                }
-            }
-        )
-
-
-class Type(str, Enum):
-    """Enumeration of the different types generated by the Firestore backend."""
-
-    NULL = "null"
-    ARRAY = "array"
-    BOOLEAN = "boolean"
-    BYTES = "bytes"
-    TIMESTAMP = "timestamp"
-    GEO_POINT = "geo_point"
-    NUMBER = "number"
-    INT32 = "int32"
-    INT64 = "int64"
-    FLOAT64 = "float64"
-    DECIMAL128 = "decimal128"
-    MAP = "map"
-    REFERENCE = "reference"
-    STRING = "string"
-    VECTOR = "vector"
-    MAX_KEY = "max_key"
-    MIN_KEY = "min_key"
-    OBJECT_ID = "object_id"
-    REGEX = "regex"
-    REQUEST_TIMESTAMP = "request_timestamp"
 
 
 class Expression(ABC):
@@ -764,6 +630,8 @@ class Expression(ABC):
         """Evaluates to the distance in meters between the location in the specified
         field and the query location.
 
+        .. note::
+            This feature is currently in beta and is subject to change.
         Note: This Expression can only be used within a `Search` stage.
 
         Example:
@@ -2074,10 +1942,6 @@ class Expression(ABC):
         `None`. The order of elements in the output array is not stable and
         shouldn't be relied upon.
 
-        This API is provided as a preview for developers and may change based
-        on feedback that we receive. Do not use this API in a production
-        environment.
-
         Example:
             >>> # Collect all values of field 'color' into an array
             >>> Field.of("color").array_agg()
@@ -2096,10 +1960,6 @@ class Expression(ABC):
         `None`. The order of elements in the output array is not stable and
         shouldn't be relied upon.
 
-        This API is provided as a preview for developers and may change based
-        on feedback that we receive. Do not use this API in a production
-        environment.
-
         Example:
             >>> # Collect distinct values of field 'color' into an array
             >>> Field.of("color").array_agg_distinct()
@@ -2114,10 +1974,6 @@ class Expression(ABC):
         """Creates an aggregation that finds the first value of an expression
         across multiple stage inputs.
 
-        This API is provided as a preview for developers and may change based
-        on feedback that we receive. Do not use this API in a production
-        environment.
-
         Example:
             >>> # Select the first value of field 'color'
             >>> Field.of("color").first()
@@ -2131,10 +1987,6 @@ class Expression(ABC):
     def last(self) -> "Expression":
         """Creates an aggregation that finds the last value of an expression
         across multiple stage inputs.
-
-        This API is provided as a preview for developers and may change based
-        on feedback that we receive. Do not use this API in a production
-        environment.
 
         Example:
             >>> # Select the last value of field 'color'
@@ -2690,7 +2542,9 @@ class Expression(ABC):
         return FunctionExpression("type", [self])
 
     @expose_as_static
-    def is_type(self, type_val: Type | str | Expression) -> "BooleanExpression":
+    def is_type(
+        self, type_val: PipelineDataType | str | Expression
+    ) -> "BooleanExpression":
         """Creates an expression that checks if the result is of the specified type.
 
         Example:
@@ -3226,6 +3080,9 @@ class Score(FunctionExpression):
     in the search query. If `SearchOptions.query` is not set or does not contain
     any text predicates, then this topicality score will always be `0`.
 
+    .. note::
+        This feature is currently in beta and is subject to change.
+
     Note: This Expression can only be used within a `Search` stage.
 
     Example:
@@ -3248,6 +3105,9 @@ class Score(FunctionExpression):
 
 class DocumentMatches(BooleanExpression):
     """Creates a boolean expression for a document match query.
+
+    .. note::
+        This feature is currently in beta and is subject to change.
 
     Note: This Expression can only be used within a `Search` stage.
