+++ packages/google-cloud-firestore/google/cloud/firestore_v1/pipeline_stages.py
@@ -11,16 +11,10 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""
-.. warning::
-    **Preview API**: Firestore Pipelines is currently in preview and is
-    subject to potential breaking changes in future releases.
-"""
 
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from enum import Enum
 from typing import TYPE_CHECKING, Any, Mapping, Optional, Sequence
 
 from google.cloud.firestore_v1._helpers import encode_value
@@ -30,10 +24,8 @@ from google.cloud.firestore_v1.pipeline_expressions import (
     AliasedExpression,
     BooleanExpression,
     CONSTANT_TYPE,
-    DocumentMatches,
     Expression,
     Field,
-    Ordering,
     Selectable,
 )
 from google.cloud.firestore_v1.types.document import Pipeline as Pipeline_pb
@@ -43,161 +35,14 @@ from google.cloud.firestore_v1.vector import Vector
 if TYPE_CHECKING:  # pragma: NO COVER
     from google.cloud.firestore_v1.base_document import BaseDocumentReference
     from google.cloud.firestore_v1.base_pipeline import _BasePipeline
+    from google.cloud.firestore_v1.pipeline_types import Ordering
 
-
-class FindNearestOptions:
-    """Options for configuring the `FindNearest` pipeline stage.
-
-    Attributes:
-        limit (Optional[int]): The maximum number of nearest neighbors to return.
-        distance_field (Optional[Field]): An optional field to store the calculated
-            distance in the output documents.
-    """
-
-    def __init__(
-        self,
-        limit: Optional[int] = None,
-        distance_field: Optional[Field] = None,
-    ):
-        self.limit = limit
-        self.distance_field = distance_field
-
-    def __repr__(self):
-        args = []
-        if self.limit is not None:
-            args.append(f"limit={self.limit}")
-        if self.distance_field is not None:
-            args.append(f"distance_field={self.distance_field}")
-        return f"{self.__class__.__name__}({', '.join(args)})"
-
-
-class SampleOptions:
-    """Options for the 'sample' pipeline stage."""
-
-    class Mode(Enum):
-        DOCUMENTS = "documents"
-        PERCENT = "percent"
-
-    def __init__(self, value: int | float, mode: Mode | str):
-        self.value = value
-        self.mode = SampleOptions.Mode[mode.upper()] if isinstance(mode, str) else mode
-
-    def __repr__(self):
-        if self.mode == SampleOptions.Mode.DOCUMENTS:
-            mode_str = "doc_limit"
-        else:
-            mode_str = "percentage"
-        return f"SampleOptions.{mode_str}({self.value})"
-
-    @staticmethod
-    def doc_limit(value: int):
-        """
-        Sample a set number of documents
-
-        Args:
-            value: number of documents to sample
-        """
-        return SampleOptions(value, mode=SampleOptions.Mode.DOCUMENTS)
-
-    @staticmethod
-    def percentage(value: float):
-        """
-        Sample a percentage of documents
-
-        Args:
-            value: percentage of documents to return
-        """
-        return SampleOptions(value, mode=SampleOptions.Mode.PERCENT)
-
-
-class SearchOptions:
-    """Options for configuring the `Search` pipeline stage."""
-
-    def __init__(
-        self,
-        query: str | BooleanExpression,
-        *,
-        limit: Optional[int] = None,
-        retrieval_depth: Optional[int] = None,
-        sort: Optional[Sequence[Ordering] | Ordering] = None,
-        add_fields: Optional[Sequence[Selectable]] = None,
-        offset: Optional[int] = None,
-        language_code: Optional[str] = None,
-    ):
-        """
-        Initializes a SearchOptions instance.
-
-        Args:
-            query (str | BooleanExpression): Specifies the search query that will be used to query and score documents
-                by the search stage. The query can be expressed as an `Expression`, which will be used to score
-                and filter the results. Not all expressions supported by Pipelines are supported in the Search query.
-                The query can also be expressed as a string in the Search DSL.
-            limit (Optional[int]): The maximum number of documents to return from the Search stage.
-            retrieval_depth (Optional[int]): The maximum number of documents for the search stage to score. Documents
-                will be processed in the pre-sort order specified by the search index.
-            sort (Optional[Sequence[Ordering] | Ordering]): Orderings specify how the input documents are sorted.
-            add_fields (Optional[Sequence[Selectable]]): The fields to add to each document, specified as a `Selectable`.
-            offset (Optional[int]): The number of documents to skip.
-            language_code (Optional[str]): The BCP-47 language code of text in the search query, such as "en-US" or "sr-Latn".
-        """
-        self.query = DocumentMatches(query) if isinstance(query, str) else query
-        self.limit = limit
-        self.retrieval_depth = retrieval_depth
-        self.sort = [sort] if isinstance(sort, Ordering) else sort
-        self.add_fields = add_fields
-        self.offset = offset
-        self.language_code = language_code
-
-    def __repr__(self):
-        args = [f"query={self.query!r}"]
-        if self.limit is not None:
-            args.append(f"limit={self.limit}")
-        if self.retrieval_depth is not None:
-            args.append(f"retrieval_depth={self.retrieval_depth}")
-        if self.sort is not None:
-            args.append(f"sort={self.sort}")
-        if self.add_fields is not None:
-            args.append(f"add_fields={self.add_fields}")
-        if self.offset is not None:
-            args.append(f"offset={self.offset}")
-        if self.language_code is not None:
-            args.append(f"language_code={self.language_code!r}")
-        return f"{self.__class__.__name__}({', '.join(args)})"
-
-    def _to_dict(self) -> dict[str, Value]:
-        options = {"query": self.query._to_pb()}
-        if self.limit is not None:
-            options["limit"] = Value(integer_value=self.limit)
-        if self.retrieval_depth is not None:
-            options["retrieval_depth"] = Value(integer_value=self.retrieval_depth)
-        if self.sort is not None:
-            options["sort"] = Value(
-                array_value={"values": [s._to_pb() for s in self.sort]}
-            )
-        if self.add_fields is not None:
-            options["add_fields"] = Selectable._to_value(self.add_fields)
-        if self.offset is not None:
-            options["offset"] = Value(integer_value=self.offset)
-        if self.language_code is not None:
-            options["language_code"] = Value(string_value=self.language_code)
-        return options
-
-
-class UnnestOptions:
-    """Options for configuring the `Unnest` pipeline stage.
-
-    Attributes:
-        index_field (str): The name of the field to add to each output document,
-            storing the original 0-based index of the element within the array.
-    """
-
-    def __init__(self, index_field: Field | str):
-        self.index_field = (
-            index_field if isinstance(index_field, Field) else Field.of(index_field)
-        )
-
-    def __repr__(self):
-        return f"{self.__class__.__name__}(index_field={self.index_field.path!r})"
+from google.cloud.firestore_v1.pipeline_types import (
+    FindNearestOptions,
+    SampleOptions,
+    SearchOptions,
+    UnnestOptions,
+)
 
 
 class Stage(ABC):
@@ -498,7 +343,11 @@ class Sample(Stage):
 
 
 class Search(Stage):
-    """Search stage."""
+    """Search stage.
+
+    .. note::
+        This feature is currently in beta and is subject to change.
+    """
 
     def __init__(self, query_or_options: str | BooleanExpression | SearchOptions):
         super().__init__("search")
@@ -589,7 +438,11 @@ class Where(Stage):
 
 
 class Delete(Stage):
-    """Deletes documents matching the pipeline criteria."""
+    """Deletes documents matching the pipeline criteria.
+
+    .. note::
+        This feature is currently in beta and is subject to change.
+    """
 
     def __init__(self):
         super().__init__("delete")
@@ -599,7 +452,11 @@ class Delete(Stage):
 
 
 class Update(Stage):
-    """Updates documents with transformed fields."""
+    """Updates documents with transformed fields.
+
+    .. note::
+        This feature is currently in beta and is subject to change.
+    """
 
     def __init__(self, *transformed_fields: Selectable):
         super().__init__("update")
