+++ packages/google-cloud-firestore/google/cloud/firestore_v1/pipeline_types.py
@@ -1,0 +1,329 @@
+# -*- coding: utf-8 -*-
+# Copyright 2026 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from __future__ import annotations
+
+from enum import Enum
+from typing import Optional, Sequence
+
+from google.cloud.firestore_v1 import pipeline_expressions
+from google.cloud.firestore_v1.types.document import Value
+
+
+class TimeUnit(str, Enum):
+    """Enumeration of the different time units supported by the Firestore backend."""
+
+    MICROSECOND = "microsecond"
+    MILLISECOND = "millisecond"
+    SECOND = "second"
+    MINUTE = "minute"
+    HOUR = "hour"
+    DAY = "day"
+
+
+class TimeGranularity(str, Enum):
+    """Enumeration of the different time granularities supported by the Firestore backend."""
+
+    # Inherit from TimeUnit
+    MICROSECOND = TimeUnit.MICROSECOND.value
+    MILLISECOND = TimeUnit.MILLISECOND.value
+    SECOND = TimeUnit.SECOND.value
+    MINUTE = TimeUnit.MINUTE.value
+    HOUR = TimeUnit.HOUR.value
+    DAY = TimeUnit.DAY.value
+
+    # Additional granularities
+    WEEK = "week"
+    WEEK_MONDAY = "week(monday)"
+    WEEK_TUESDAY = "week(tuesday)"
+    WEEK_WEDNESDAY = "week(wednesday)"
+    WEEK_THURSDAY = "week(thursday)"
+    WEEK_FRIDAY = "week(friday)"
+    WEEK_SATURDAY = "week(saturday)"
+    WEEK_SUNDAY = "week(sunday)"
+    ISOWEEK = "isoweek"
+    MONTH = "month"
+    QUARTER = "quarter"
+    YEAR = "year"
+    ISOYEAR = "isoyear"
+
+
+class TimePart(str, Enum):
+    """Enumeration of the different time parts supported by the Firestore backend."""
+
+    # Inherit from TimeUnit
+    MICROSECOND = TimeUnit.MICROSECOND.value
+    MILLISECOND = TimeUnit.MILLISECOND.value
+    SECOND = TimeUnit.SECOND.value
+    MINUTE = TimeUnit.MINUTE.value
+    HOUR = TimeUnit.HOUR.value
+    DAY = TimeUnit.DAY.value
+
+    # Inherit from TimeGranularity
+    WEEK = TimeGranularity.WEEK.value
+    WEEK_MONDAY = TimeGranularity.WEEK_MONDAY.value
+    WEEK_TUESDAY = TimeGranularity.WEEK_TUESDAY.value
+    WEEK_WEDNESDAY = TimeGranularity.WEEK_WEDNESDAY.value
+    WEEK_THURSDAY = TimeGranularity.WEEK_THURSDAY.value
+    WEEK_FRIDAY = TimeGranularity.WEEK_FRIDAY.value
+    WEEK_SATURDAY = TimeGranularity.WEEK_SATURDAY.value
+    WEEK_SUNDAY = TimeGranularity.WEEK_SUNDAY.value
+    ISOWEEK = TimeGranularity.ISOWEEK.value
+    MONTH = TimeGranularity.MONTH.value
+    QUARTER = TimeGranularity.QUARTER.value
+    YEAR = TimeGranularity.YEAR.value
+    ISOYEAR = TimeGranularity.ISOYEAR.value
+
+    # Additional parts
+    DAY_OF_WEEK = "dayofweek"
+    DAY_OF_YEAR = "dayofyear"
+
+
+class PipelineDataType(str, Enum):
+    """Enumeration of the different types generated by the Firestore backend."""
+
+    NULL = "null"
+    ARRAY = "array"
+    BOOLEAN = "boolean"
+    BYTES = "bytes"
+    TIMESTAMP = "timestamp"
+    GEO_POINT = "geo_point"
+    NUMBER = "number"
+    INT32 = "int32"
+    INT64 = "int64"
+    FLOAT64 = "float64"
+    DECIMAL128 = "decimal128"
+    MAP = "map"
+    REFERENCE = "reference"
+    STRING = "string"
+    VECTOR = "vector"
+    MAX_KEY = "max_key"
+    MIN_KEY = "min_key"
+    OBJECT_ID = "object_id"
+    REGEX = "regex"
+    REQUEST_TIMESTAMP = "request_timestamp"
+
+
+class Ordering:
+    """Represents the direction for sorting results in a pipeline."""
+
+    class Direction(Enum):
+        ASCENDING = "ascending"
+        DESCENDING = "descending"
+
+    def __init__(self, expr, order_dir: Direction | str = Direction.ASCENDING):
+        """
+        Initializes an Ordering instance
+
+        Args:
+            expr (Expression | str): The expression or field path string to sort by.
+                If a string is provided, it's treated as a field path.
+            order_dir (Direction | str): The direction to sort in.
+                Defaults to ascending
+        """
+        self.expr = (
+            expr
+            if isinstance(expr, pipeline_expressions.Expression)
+            else pipeline_expressions.Field.of(expr)
+        )
+        self.order_dir = (
+            Ordering.Direction[order_dir.upper()]
+            if isinstance(order_dir, str)
+            else order_dir
+        )
+
+    def __repr__(self):
+        if self.order_dir is Ordering.Direction.ASCENDING:
+            order_str = ".ascending()"
+        else:
+            order_str = ".descending()"
+        return f"{self.expr!r}{order_str}"
+
+    def _to_pb(self) -> Value:
+        return Value(
+            map_value={
+                "fields": {
+                    "direction": Value(string_value=self.order_dir.value),
+                    "expression": self.expr._to_pb(),
+                }
+            }
+        )
+
+
+class FindNearestOptions:
+    """Options for configuring the `FindNearest` pipeline stage.
+
+    Attributes:
+        limit (Optional[int]): The maximum number of nearest neighbors to return.
+        distance_field (Optional[Field]): An optional field to store the calculated
+            distance in the output documents.
+    """
+
+    def __init__(
+        self,
+        limit: Optional[int] = None,
+        distance_field: Optional[pipeline_expressions.Field] = None,
+    ):
+        self.limit = limit
+        self.distance_field = distance_field
+
+    def __repr__(self):
+        args = []
+        if self.limit is not None:
+            args.append(f"limit={self.limit}")
+        if self.distance_field is not None:
+            args.append(f"distance_field={self.distance_field}")
+        return f"{self.__class__.__name__}({', '.join(args)})"
+
+
+class SampleOptions:
+    """Options for the 'sample' pipeline stage."""
+
+    class Mode(Enum):
+        DOCUMENTS = "documents"
+        PERCENT = "percent"
+
+    def __init__(self, value: int | float, mode: Mode | str):
+        self.value = value
+        self.mode = SampleOptions.Mode[mode.upper()] if isinstance(mode, str) else mode
+
+    def __repr__(self):
+        if self.mode == SampleOptions.Mode.DOCUMENTS:
+            mode_str = "doc_limit"
+        else:
+            mode_str = "percentage"
+        return f"SampleOptions.{mode_str}({self.value})"
+
+    @staticmethod
+    def doc_limit(value: int):
+        """
+        Sample a set number of documents
+
+        Args:
+            value: number of documents to sample
+        """
+        return SampleOptions(value, mode=SampleOptions.Mode.DOCUMENTS)
+
+    @staticmethod
+    def percentage(value: float):
+        """
+        Sample a percentage of documents
+
+        Args:
+            value: percentage of documents to return
+        """
+        return SampleOptions(value, mode=SampleOptions.Mode.PERCENT)
+
+
+class SearchOptions:
+    """Options for configuring the `Search` pipeline stage.
+
+    .. note::
+        This feature is currently in beta and is subject to change.
+    """
+
+    def __init__(
+        self,
+        query: str | pipeline_expressions.BooleanExpression,
+        *,
+        limit: Optional[int] = None,
+        retrieval_depth: Optional[int] = None,
+        sort: Optional[Sequence[Ordering] | Ordering] = None,
+        add_fields: Optional[Sequence[pipeline_expressions.Selectable]] = None,
+        offset: Optional[int] = None,
+        language_code: Optional[str] = None,
+    ):
+        """
+        Initializes a SearchOptions instance.
+
+        Args:
+            query: Specifies the search query that will be used to query and score documents
+                by the search stage. The query can be expressed as an `Expression`, which will be used to score
+                and filter the results. Not all expressions supported by Pipelines are supported in the Search query.
+                The query can also be expressed as a string in the Search DSL.
+            limit: The maximum number of documents to return from the Search stage.
+            retrieval_depth: The maximum number of documents for the search stage to score. Documents
+                will be processed in the pre-sort order specified by the search index.
+            sort: Orderings specify how the input documents are sorted.
+            add_fields: The fields to add to each document, specified as a `Selectable`.
+            offset: The number of documents to skip.
+            language_code: The BCP-47 language code of text in the search query, such as "en-US" or "sr-Latn".
+        """
+        self.query = (
+            pipeline_expressions.DocumentMatches(query)
+            if isinstance(query, str)
+            else query
+        )
+        self.limit = limit
+        self.retrieval_depth = retrieval_depth
+        self.sort = [sort] if isinstance(sort, Ordering) else sort
+        self.add_fields = add_fields
+        self.offset = offset
+        self.language_code = language_code
+
+    def __repr__(self):
+        args = [f"query={self.query!r}"]
+        if self.limit is not None:
+            args.append(f"limit={self.limit}")
+        if self.retrieval_depth is not None:
+            args.append(f"retrieval_depth={self.retrieval_depth}")
+        if self.sort is not None:
+            args.append(f"sort={self.sort}")
+        if self.add_fields is not None:
+            args.append(f"add_fields={self.add_fields}")
+        if self.offset is not None:
+            args.append(f"offset={self.offset}")
+        if self.language_code is not None:
+            args.append(f"language_code={self.language_code!r}")
+        return f"{self.__class__.__name__}({', '.join(args)})"
+
+    def _to_dict(self) -> dict[str, Value]:
+        options = {"query": self.query._to_pb()}
+        if self.limit is not None:
+            options["limit"] = Value(integer_value=self.limit)
+        if self.retrieval_depth is not None:
+            options["retrieval_depth"] = Value(integer_value=self.retrieval_depth)
+        if self.sort is not None:
+            options["sort"] = Value(
+                array_value={"values": [s._to_pb() for s in self.sort]}
+            )
+        if self.add_fields is not None:
+            options["add_fields"] = pipeline_expressions.Selectable._to_value(
+                self.add_fields
+            )
+        if self.offset is not None:
+            options["offset"] = Value(integer_value=self.offset)
+        if self.language_code is not None:
+            options["language_code"] = Value(string_value=self.language_code)
+        return options
+
+
+class UnnestOptions:
+    """Options for configuring the `Unnest` pipeline stage.
+
+    Attributes:
+        index_field (str): The name of the field to add to each output document,
+            storing the original 0-based index of the element within the array.
+    """
+
+    def __init__(self, index_field: pipeline_expressions.Field | str):
+        self.index_field = (
+            index_field
+            if isinstance(index_field, pipeline_expressions.Field)
+            else pipeline_expressions.Field.of(index_field)
+        )
+
+    def __repr__(self):
+        return f"{self.__class__.__name__}(index_field={self.index_field.path!r})"
