+++ packages/google-cloud-firestore/tests/unit/v1/test_async_pipeline.py
@@ -450,6 +450,19 @@ def test_async_pipeline_aggregate_with_groups():
     assert list(result_ppl.stages[0].accumulators) == [Field.of("title")]
 
 
+def test_async_pipeline_raw_stage_with_options():
+    from google.cloud.firestore_v1.pipeline_stages import RawStage
+
+    start_ppl = _make_async_pipeline()
+    result_ppl = start_ppl.raw_stage(
+        "stage_name", Field.of("n"), options={"key": "val"}
+    )
+    assert len(start_ppl.stages) == 0
+    assert len(result_ppl.stages) == 1
+    assert isinstance(result_ppl.stages[0], RawStage)
+    assert result_ppl.stages[0].options == {"key": "val"}
+
+
 def test_async_pipeline_union_relative_error():
     start_ppl = _make_async_pipeline(client=mock.Mock())
     other_ppl = _make_async_pipeline(client=None)
